springboot-sample/src/main/java/com/abaghel/examples/spring/springboot/hikaricp/HikariCPConn.java
package com.abaghel.examples.spring.springboot.hikaricp;

import javax.sql.DataSource;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.context.annotation.Bean;
import org.springframework.jdbc.core.JdbcTemplate;
import org.springframework.stereotype.Component;
/**
 * 
 * @author abaghel
 *
 */
@Component
public class HikariCPConn {
	
	@Autowired
	DataSource dataSource;

	@Bean
	public JdbcTemplate getJdbcTemplate() {
		return new JdbcTemplate(dataSource);
	}
}
